adds stubwasi.rs
